re-name directory
