+++ .env
@@ -1,0 +1,4 @@
+DB_HOST=34.134.121.66
+DB_USER=root
+DB_PASS=123
+DB_NAME=workout_app_backend
