+++ package.json
@@ -18,6 +18,7 @@
     "util": "^0.12.4"
   },
   "devDependencies": {
+    "dotenv": "^10.0.0",
     "nodemon": "^2.0.13",
     "pretier": "0.0.1"
   }
