+++ src/controllers/login.js
@@ -1,0 +1,28 @@
+const LoginModel = require("../models/login");
+
+module.exports = {
+  signUp: async (req, res) => {
+    try {
+      const username = req.body.username;
+      const password = req.body.password;
+
+      const result = await LoginModel.signUp(username, password);
+
+      res.status(201).json(result);
+    } catch (e) {
+      res.status(500).send({ e });
+    }
+  },
+
+  signIn: async (req, res) => {
+    try {
+      const username = req.body.username;
+      const password = req.body.password;
+      const result = await LoginModel.signIn(username, password);
+
+      res.status(200).json(result);
+    } catch (e) {
+      res.status(500).send({ e });
+    }
+  },
+};
