+++ src/models/workout.js
@@ -1,49 +1,12 @@
 const util = require("util");
-const bcrypt = require("bcrypt");
-const connection = require("./db");
+const connection = require("../db");
 
 const WorkoutModel = {
-  signIn: async (username, password) => {
-    try {
-      const hashed_password = bcrypt.hashSync(password, 10);
-
-      const resultSignIn = await util
-        .promisify(connection.execute)
-        .bind(connection)(
-        `INSERT INTO user (username, password) VALUES ('${username}', '${hashed_password}')`
-      );
-
-      return resultSignIn;
-    } catch (e) {
-      throw new Error(e);
-    }
-  },
-
-  login: async (username, password) => {
-    try {
-      const resultFindUsername = await util
-        .promisify(connection.execute)
-        .bind(connection)(`SELECT * FROM user WHERE username = '${username}'`);
-      if (resultFindUsername.length <= 0) {
-        return "NO USERS FOUND";
-      }
-
-      const isValidPass = await verifyPass(
-        password,
-        resultFindUsername[0].password
-      );
-
-      return { id: resultFindUsername[0].id, isValidLogin: isValidPass };
-    } catch (e) {
-      throw new Error(e);
-    }
-  },
-
-  createUserWorkouts: async (day, userId) => {
+  createUserWorkouts: async (dayOfWeek, userId) => {
     try {
       let queryState;
-      if (day !== null) {
-        queryState = `INSERT INTO user_workouts (scheduled_day, user_id) VALUES('${day}', '${userId}')`;
+      if (dayOfWeek !== null) {
+        queryState = `INSERT INTO user_workouts (scheduled_day, user_id) VALUES('${dayOfWeek}', '${userId}')`;
       } else {
         queryState = `INSERT INTO user_workouts (user_id) VALUES(${userId})`;
       }
@@ -112,7 +75,7 @@ const WorkoutModel = {
     }
   },
   updateWorkoutWithDayChange: async ({
-    newDay,
+    newDayOfWeek,
     workoutId,
     userId,
     category,
@@ -126,13 +89,11 @@ const WorkoutModel = {
         .bind(connection)(
         `SELECT max(order_index) as maxOrderIndex FROM workout_sets 
         LEFT JOIN user_workouts ON user_workouts.id = user_workouts_id 
-        WHERE workout_sets.user_workouts_user_id = ${userId} and scheduled_day="${newDay}";`
+        WHERE workout_sets.user_workouts_user_id = ${userId} and scheduled_day="${newDayOfWeek}";`
       );
-
-      maxOrderIndex =
-        maxOrderIndex[0].maxOrderIndex !== null
-          ? maxOrderIndex[0].maxOrderIndex + 1
-          : 1;
+      maxOrderIndex = maxOrderIndex[0].maxOrderIndex
+        ? maxOrderIndex[0].maxOrderIndex + 1
+        : 0;
 
       const resultUpdateWorkoutWithDayChange = await util
         .promisify(connection.execute)
@@ -142,7 +103,7 @@ const WorkoutModel = {
         LEFT JOIN user_workouts as UW ON UW.id = user_workouts_id 
         SET training_name=?, sets=?, times=?, order_index=?, label=?, scheduled_day=?
         WHERE user_workouts_id= ?`,
-        [name, sets, times, maxOrderIndex, category, newDay, workoutId]
+        [name, sets, times, maxOrderIndex, category, newDayOfWeek, workoutId]
       );
 
       return resultUpdateWorkoutWithDayChange;
@@ -165,38 +126,6 @@ const WorkoutModel = {
       throw new Error(e);
     }
   },
-
-  // updateSchedule: async (newDay, userId) => {
-  //   try {
-  //     const resultUpdateSchedule = await util
-  //       .promisify(connection.execute)
-  //       .bind(connection)(`UPDATE user_workouts SET scheduled_day = ${newDay}`);
-  //     return createWorkoutSet;
-  //   } catch (e) {}
-  // },
-
-  getUserScheduleInfo: async (id) => {
-    try {
-      const getUserScheduleInfo = await util
-        .promisify(connection.execute)
-        .bind(connection)(
-        `SELECT workout_sets.user_workouts_user_id as userId, user_workouts_id ,workout_sets.training_name, workout_sets.sets, workout_sets.times, workout_sets.order_index, workout_categories.label, user_workouts.scheduled_day FROM workout_app_backend.workout_sets 
-        LEFT JOIN workout_categories ON workout_categories.id=workout_categories_id 
-        LEFT JOIN user_workouts ON user_workouts.id = user_workouts_id 
-        WHERE workout_sets.user_workouts_user_id = ${id};`
-      );
-
-      return getUserScheduleInfo;
-    } catch (e) {
-      throw new Error(e);
-    }
-  },
 };
-
-async function verifyPass(passA, passB) {
-  const passwordCheck = await bcrypt.compare(passA, passB);
-
-  return passwordCheck;
-}
 
 module.exports = WorkoutModel;
