+++ src/server.js
@@ -1,6 +1,6 @@
+require("dotenv").config();
 const express = require("express");
 const app = express();
-
 const router = require("./router");
 
 app.use(router);
